bot/requirements.txt
python-telegram-bot==21.0
aider-chat
requests
python-dotenv
